refactor: 지원서 관련 apis 함수들 apis/application.ts로 이동
